+++ test/discrete/test_rainbow.py
@@ -8,7 +8,7 @@ import numpy as np
 from torch.utils.tensorboard import SummaryWriter
 
 from tianshou.policy import RainbowPolicy
-from tianshou.utils import BasicLogger
+from tianshou.utils import TensorboardLogger
 from tianshou.env import DummyVectorEnv
 from tianshou.utils.net.common import Net
 from tianshou.utils.net.discrete import NoisyLinear
@@ -102,7 +102,7 @@ def test_rainbow(args=get_args()):
     # log
     log_path = os.path.join(args.logdir, args.task, 'rainbow')
     writer = SummaryWriter(log_path)
-    logger = BasicLogger(writer, save_interval=args.save_interval)
+    logger = TensorboardLogger(writer, save_interval=args.save_interval)
 
     def save_fn(policy):
         torch.save(policy.state_dict(), os.path.join(log_path, 'policy.pth'))
