+++ test/multiagent/Gomoku.py
@@ -7,7 +7,7 @@ from torch.utils.tensorboard import SummaryWriter
 from tianshou.env import DummyVectorEnv
 from tianshou.data import Collector
 from tianshou.policy import RandomPolicy
-from tianshou.utils import BasicLogger
+from tianshou.utils import TensorboardLogger
 
 from tic_tac_toe_env import TicTacToeEnv
 from tic_tac_toe import get_parser, get_agents, train_agent, watch
@@ -33,7 +33,7 @@ def gomoku(args=get_args()):
     # log
     log_path = os.path.join(args.logdir, 'Gomoku', 'dqn')
     writer = SummaryWriter(log_path)
-    args.logger = BasicLogger(writer)
+    args.logger = TensorboardLogger(writer)
 
     opponent_pool = [agent_opponent]
