+++ tianshou/utils/__init__.py
@@ -1,12 +1,17 @@
 from tianshou.utils.config import tqdm_config
 from tianshou.utils.statistics import MovAvg, RunningMeanStd
-from tianshou.utils.log_tools import BasicLogger, LazyLogger, BaseLogger
+from tianshou.utils.logger.base import BaseLogger, LazyLogger
+from tianshou.utils.logger.tensorboard import TensorboardLogger, BasicLogger
+from tianshou.utils.logger.wandb import WandBLogger
+
 
 __all__ = [
     "MovAvg",
     "RunningMeanStd",
     "tqdm_config",
     "BaseLogger",
+    "TensorboardLogger",
     "BasicLogger",
     "LazyLogger",
+    "WandBLogger"
 ]
